Extract shared WinPE workspace preparation from media output
