+++ src/Foundry/Services/WinPe/IWinPeWorkspacePreparationService.cs
@@ -1,0 +1,9 @@
+namespace Foundry.Services.WinPe;
+
+internal interface IWinPeWorkspacePreparationService
+{
+    Task<WinPeResult<WinPeWorkspacePreparationResult>> PrepareAsync(
+        WinPeWorkspacePreparationRequest request,
+        IProgress<WinPeWorkspacePreparationStage>? progress,
+        CancellationToken cancellationToken);
+}
