+++ src/Foundry/Services/WinPe/WinPeWorkspacePreparationResult.cs
@@ -1,0 +1,6 @@
+namespace Foundry.Services.WinPe;
+
+internal sealed record WinPeWorkspacePreparationResult
+{
+    public bool UseBootEx { get; init; }
+}
