+++ src/Foundry/Services/WinPe/WinPeWorkspacePreparationStage.cs
@@ -1,0 +1,8 @@
+namespace Foundry.Services.WinPe;
+
+internal enum WinPeWorkspacePreparationStage
+{
+    ResolvingDrivers,
+    CustomizingImage,
+    EvaluatingSignaturePolicy
+}
